refactor: MetadataViewer page
